// app/profile/types.ts
export interface ProfileData {
  id?: string;
  username?: string;
  full_name?: string;
  avatar_url?: string;
  title_prefix?: string;
  first_name?: string;
  last_name?: string;
  employee_id?: string;
  position?: string;
  department?: string;
  work_location?: string;
  work_latitude?: number;
  work_longitude?: number;
  work_radius?: number;
  updated_at?: string;
}